Move large classes to their own files
